♻️ 모바일 관련 리팩토링
